+++ CHANGELOG.md
@@ -53,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow the `Attribute` suffix to be omitted for attribute references in `MixedNames`.
 - Allow the extended type name to be anywhere in the helper name (rather than requiring it at the
   start) in `HelperName`.
+- Take a fully-qualified type name for the parent type in `InheritedTypeName`, instead of a regular
+  expression.
 - Improve name resolution for declarations within types.
 - Improve type resolution for array accesses into variants.
 - Improve parsing and type modeling around `AnsiString` types with specified code pages.